Optional and variable length random nonces
The random nonces are only needed when the ephemeral keys are reused, an curve25519 publik key is a 32 bytes uniformly random byte string.

The fixed length of 64 bits puts a limit on the security level that can be achieved when ephemeral keys are reused. I.e. when ephemeral keys are reused, the security level for the aliveness of the other endpoint is only 64 bits. The security level of this should be adjustable like the security level of the AEAD, HKDF, Signature etc.