Session ID -> Connection ID
The term Session ID was causing some confusion from the people doing formal verification. EDHOC does not use the session identifiers as in [SIGMA]. I therefore suggest changing the name to Connection ID. This is the term that DTLS use for similar functionality and should therefore be a good name to avoid confusion.