index in SUITES_I not needed anymore
After the option to signal a cipher suite as an array of algorithms has been removed the uint index is in many cases not reducing any bytes and instead a int suite can be used for the selected cipher suite. This should hopefully make things easier as there has been some questions regarding the use of the uint index.